requirements.txt
email-validator==2.2.0
flask>=3.0.0
Flask==3.1.0
Flask-Cors==5.0.0
paramiko>=3.4.0
psycopg2-binary>=2.9.9
psycopg[binary]==3.2.12
PyJWT==2.10.1
python-dotenv>=1.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
# Web Framework
Flask
flask-cors

# Database
psycopg2-binary

# AWS & Auth
boto3
python-dotenv

# Inter-service Communication 
requests

#etc
uuid
datetime